+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version [4.0.0](https://github.com/cedx/where.dart/compare/v3.0.0...v4.0.0)
 - Breaking change: starting to use and support the [Dart 2 SDK](https://www.dartlang.org/tools/sdk).
+- Added the `FinderException` class.
 - Updated the package dependencies.
 
 ## Version [3.0.0](https://github.com/cedx/where.dart/compare/v2.2.0...v3.0.0)